requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
eval-type-backport==0.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
markdown==3.8
mistralai==1.7.0
pip==25.1.1
pydantic==2.11.4
pydantic-core==2.33.2
pyside6==6.9.0
pyside6-addons==6.9.0
pyside6-essentials==6.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
shiboken6==6.9.0
six==1.17.0
sniffio==1.3.1
typing-extensions==4.13.2
typing-inspection==0.4.0
uv==0.7.5
uvloop==0.21.0
